backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
scikit-image==0.22.0
scipy==1.11.4
torch==2.1.0
torchvision==0.16.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
cors==1.0.1
aiofiles==23.2.1

# Novel algorithm dependencies
PyWavelets==1.4.1
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
# GDAL==3.8.2  # Commented out due to build issues
rasterio==1.3.9
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1

# Statistical analysis
statsmodels==0.14.0
pandas==2.1.4
plotly==5.17.0

# Image processing enhancements
imageio==2.31.5
mahotas==1.4.13

# Geospatial processing
earthpy==0.9.4
folium==0.15.1 